move deployment logic to deployer pkg
