+++ .goreleaser.yaml
@@ -45,8 +45,8 @@ builds:
     ignore:
       - goos: windows
     ldflags:
-      - -X github.com/threefoldtech/tfgrid-sdk-go/grid-cli/cmd.version={{.Tag}}
-      - -X github.com/threefoldtech/tfgrid-sdk-go/grid-cli/cmd.commit={{.Commit}}
+      - -X github.com/threefoldtech/tfgrid-sdk-go/mass-deployer/cmd.version={{.Tag}}
+      - -X github.com/threefoldtech/tfgrid-sdk-go/mass-deployer/cmd.commit={{.Commit}}
 
   - dir: ./grid-proxy/cmds/proxy_server
     env:
